requirements.txt
facenet_pytorch==2.6.0
torch
numpy==1.17.0
tqdm==4.31.1
torchvision==0.17.2
opencv_python==4.12.0.88
Pillow==10.2.0